log grpc server message dispatch, prefix server log entries with hostname if logfile unspecified
